+++ r4300/bc.c
@@ -49,7 +49,7 @@ void BC1F()
 void BC1F_OUT()
 {
    if (check_cop1_unusable()) return;
-   jump_target = (long)PC->f.i.immediate;
+   jump_target = (s32)PC->f.i.immediate;
    PC++;
    delay_slot=1;
    PC->ops();
@@ -63,7 +63,8 @@ void BC1F_OUT()
 
 void BC1F_IDLE()
 {
-   long skip;
+   s32 skip;
+
    if ((FCR31 & 0x800000)==0)
      {
 	update_count();
@@ -91,7 +92,7 @@ void BC1T()
 void BC1T_OUT()
 {
    if (check_cop1_unusable()) return;
-   jump_target = (long)PC->f.i.immediate;
+   jump_target = (s32)PC->f.i.immediate;
    PC++;
    delay_slot=1;
    PC->ops();
@@ -105,7 +106,8 @@ void BC1T_OUT()
 
 void BC1T_IDLE()
 {
-   long skip;
+   s32 skip;
+
    if ((FCR31 & 0x800000)!=0)
      {
 	update_count();
@@ -140,7 +142,7 @@ void BC1FL_OUT()
    if (check_cop1_unusable()) return;
    if ((FCR31 & 0x800000)==0)
      {
-	jump_target = (long)PC->f.i.immediate;
+	jump_target = (s32)PC->f.i.immediate;
 	PC++;
 	delay_slot=1;
 	PC->ops();
@@ -192,7 +194,7 @@ void BC1TL_OUT()
    if (check_cop1_unusable()) return;
    if ((FCR31 & 0x800000)!=0)
      {
-	jump_target = (long)PC->f.i.immediate;
+	jump_target = (s32)PC->f.i.immediate;
 	PC++;
 	delay_slot=1;
 	PC->ops();
@@ -209,7 +211,8 @@ void BC1TL_OUT()
 
 void BC1TL_IDLE()
 {
-   long skip;
+   s32 skip;
+
    if ((FCR31 & 0x800000)!=0)
      {
 	update_count();
