+++ r4300/cop1_d.c
@@ -105,7 +105,7 @@ void ROUND_L_D()
 {
    if (check_cop1_unusable()) return;
    set_round();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
+   *((s64*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -113,7 +113,7 @@ void TRUNC_L_D()
 {
    if (check_cop1_unusable()) return;
    set_trunc();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
+   *((s64*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -121,7 +121,7 @@ void CEIL_L_D()
 {
    if (check_cop1_unusable()) return;
    set_ceil();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
+   *((s64*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -129,7 +129,7 @@ void FLOOR_L_D()
 {
    if (check_cop1_unusable()) return;
    set_floor();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
+   *((s64*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -137,7 +137,7 @@ void ROUND_W_D()
 {
    if (check_cop1_unusable()) return;
    set_round();
-   *((long*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
+   *((s32*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -145,7 +145,7 @@ void TRUNC_W_D()
 {
    if (check_cop1_unusable()) return;
    set_trunc();
-   *((long*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
+   *((s32*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -153,7 +153,7 @@ void CEIL_W_D()
 {
    if (check_cop1_unusable()) return;
    set_ceil();
-   *((long*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
+   *((s32*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -161,7 +161,7 @@ void FLOOR_W_D()
 {
    if (check_cop1_unusable()) return;
    set_floor();
-   *((long*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
+   *((s32*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -177,7 +177,7 @@ void CVT_W_D()
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *((long*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
+   *((s32*)reg_cop1_simple[cffd]) = *reg_cop1_double[cffs];
    PC++;
 }
 
@@ -185,7 +185,7 @@ void CVT_L_D()
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
+   *((s64*)(reg_cop1_double[cffd])) = *reg_cop1_double[cffs];
    PC++;
 }
