+++ r4300/recomph.h
@@ -37,7 +37,7 @@ extern int max_code_length;
 extern unsigned char **inst_pointer;
 extern precomp_block* dst_block;
 extern int jump_marker;
-extern unsigned long *return_address;
+extern u32 *return_address;
 extern int fast_memory;
 
 void passe2(precomp_instr *dest, int start, int end, precomp_block* block);
@@ -45,7 +45,7 @@ void init_assembler(void *block_jumps_table, int block_jumps_number);
 void free_assembler(void **block_jumps_table, int *block_jumps_number);
 void stop_it();
 
-void gencallinterp(unsigned long addr, int jump);
+void gencallinterp(u32 addr, int jump);
 
 void genupdate_system(int type);
 void genbnel();
